Extract DefaultCacheCleanupExecutor
